# apps/api/swagger.json
{"openapi":"3.0.3","info":{"title":"Sportora API","version":"1.0.0","description":"AI Powered Sports Tournament Ecosystem API"},"servers":[{"url":"http://localhost:5000/api/v1","description":"Local Development"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"paths":{"/users/me":{"get":{"summary":"Get current logged in user","tags":["Users"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Current user profile"},"401":{"description":"Unauthorized"}}}},"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new user","description":"Creates a new Sportora user account.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["fullName","email","password"],"properties":{"fullName":{"type":"string","example":"Utkarsh Tripathi"},"email":{"type":"string","example":"utkarsh@gmail.com"},"password":{"type":"string","example":"Password123"},"phone":{"type":"string","example":"9876543210"},"role":{"type":"string","enum":["PLAYER","ORGANIZER","ADMIN","REFEREE","UMPIRE","VOLUNTEER","SPONSOR"],"example":"PLAYER"}}}}}},"responses":{"201":{"description":"User registered successfully"}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login user","description":"Authenticate a user and return JWT tokens.","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","example":"utkarsh@gmail.com"},"password":{"type":"string","example":"Password123"}}}}}},"responses":{"200":{"description":"Login successful"},"401":{"description":"Invalid credentials"}}}},"/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh Access Token","description":"Generate a new access token using the refresh token stored in the HTTP-only cookie.","responses":{"200":{"description":"Access token refreshed successfully"},"401":{"description":"Refresh token missing or invalid"}}}},"/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout user","description":"Clears refresh token cookie.","responses":{"200":{"description":"Logout successful"}}}},"/tournament/create":{"post":{"tags":["Tournament"],"summary":"Create Tournament","description":"Only ADMIN and ORGANIZER can create tournaments.","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"201":{"description":"Tournament Created"}}}},"/tournament":{"get":{"tags":["Tournament"],"summary":"Get all tournaments","description":"Get all tournaments with pagination, search and filters.","parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"example":1},{"in":"query","name":"limit","schema":{"type":"integer"},"example":10},{"in":"query","name":"search","schema":{"type":"string"},"example":"Cricket"},{"in":"query","name":"city","schema":{"type":"string"},"example":"Jaipur"},{"in":"query","name":"sport","schema":{"type":"string"},"example":"Football"}],"responses":{"200":{"description":"Tournament list fetched successfully"}}}},"/tournament/{id}":{"get":{"tags":["Tournament"],"summary":"Get Tournament by ID","description":"Returns a single tournament.","parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"example":"6a6ed53034465de88d4c289e"}],"responses":{"200":{"description":"Tournament fetched successfully"},"404":{"description":"Tournament not found"}}},"patch":{"tags":["Tournament"],"summary":"Update Tournament","description":"Update a tournament by ID.","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"example":"6a6ed53034465de88d4c289e"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object"}}}},"responses":{"200":{"description":"Tournament updated successfully"},"403":{"description":"Forbidden"},"404":{"description":"Tournament not found"}}},"delete":{"tags":["Tournament"],"summary":"Delete Tournament","description":"Delete a tournament by ID.","security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"example":"6a6ed53034465de88d4c289e"}],"responses":{"200":{"description":"Tournament deleted successfully"},"403":{"description":"Forbidden"},"404":{"description":"Tournament not found"}}}}},"tags":[]}